+++ core/src/main/scala/ai/lum/odinson/ExtractorEngine.scala
@@ -21,7 +21,7 @@ import ai.lum.odinson.lucene.analysis.TokenStreamUtils
 import ai.lum.odinson.lucene.search._
 import ai.lum.odinson.state.{ MockState, State }
 import ai.lum.odinson.digraph.Vocabulary
-import ai.lum.odinson.utils.MostRecentlyUsed
+import ai.lum.odinson.utils.{ IndexSettings, MostRecentlyUsed }
 import ai.lum.odinson.utils.exceptions.OdinsonException
 
 import scala.collection.mutable.ArrayBuffer
@@ -30,6 +30,7 @@ class ExtractorEngine private (
   val indexSearcher: OdinsonIndexSearcher,
   val compiler: QueryCompiler,
   val displayField: String,
+  val indexSettings: IndexSettings,
   val state: State, // todo: should this be private?
   val parentDocIdField: String
 ) {
@@ -104,7 +105,7 @@ class ExtractorEngine private (
     * will be returned, instead of just the correct one according to the query semantics,
     * e.g., select the longest match for the greedy quantifiers.
     *
-    * If you don't know why you should disable the MatchSelector, then keeep it enabled.
+    * If you don't know why you should disable the MatchSelector, then keep it enabled.
     *
     * @param odinsonQuery
     * @param disableMatchSelector
@@ -629,6 +630,7 @@ object ExtractorEngine {
     val displayField = config[String]("odinson.displayField")
     val indexSearcher = new OdinsonIndexSearcher(indexReader, computeTotalHits)
     val vocabulary = Vocabulary.fromDirectory(indexDir)
+    val indexSettings = IndexSettings.fromDirectory(indexDir)
     val compiler = QueryCompiler(config, vocabulary)
     val state = State(config, indexSearcher)
     val parentDocIdField = config[String]("odinson.index.documentIdField")
@@ -636,6 +638,7 @@ object ExtractorEngine {
       indexSearcher,
       compiler,
       displayField,
+      indexSettings,
       state,
       parentDocIdField
     )