+++ core/src/main/scala/ai/lum/odinson/OdinsonIndexWriter.scala
@@ -23,6 +23,7 @@ import ai.lum.odinson.lucene.analysis._
 import ai.lum.odinson.digraph.{ DirectedGraph, Vocabulary }
 import ai.lum.odinson.serialization.UnsafeSerializer
 import ai.lum.odinson.utils.IndexSettings
+import ai.lum.odinson.utils.exceptions.OdinsonException
 
 class OdinsonIndexWriter(
   val directory: Directory,
@@ -284,6 +285,12 @@ object OdinsonIndexWriter {
         val vocab = Vocabulary.fromDirectory(dir)
         (dir, vocab)
     }
+
+    // Always store the display field, also store these additional fields
+    if (!storedFields.contains(displayField)) {
+      throw new OdinsonException("`odinson.index.storedFields` must contain `odinson.displayField`")
+    }
+    val settings = IndexSettings(storedFields)
     new OdinsonIndexWriter(
       // format: off
       directory            = directory,